+++ kicost/__main__.py
@@ -1,6 +1,6 @@
 # MIT license
 #
-# Copyright (C) 2015 by XESS Corporation
+# Copyright (C) 2018 by XESS Corporation / Hildo G Jr
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -29,13 +29,21 @@ from builtins import open
 from future import standard_library
 standard_library.install_aliases()
 
-import argparse as ap
+import argparse as ap # Command argument parser.
 import os
 import sys
 import logging
 import time
-from .kicost import *
-from . import __version__
+#import inspect # To get the internal module and informations of a module/class.
+from .kicost import kicost # kicost core functions.
+try:
+    from .kicost_gui import * # User guide.
+except ImportError:
+    pass # If the wxPython dependences are not installed and
+         # the user just want the KiCost CLI.
+from .distributors import distributors
+from .eda_tools import eda_tool #from . import eda_tools as eda_tools_imports
+from . import __version__ # Version control by @xesscorp.
 
 NUM_PROCESSES = 30  # Maximum number of parallel web-scraping processes..
 HTML_RESPONSE_RETRIES = 2 # Number of attempts to retrieve part data from a website.
@@ -112,6 +120,9 @@ def main():
     parser.add_argument('--show_dist_list',
                         action='store_true',
                         help='Show list of distributors that can be scraped for cost data, then exit.')
+    parser.add_argument('--show_eda_list',
+                        action='store_true',
+                        help='Show list of eda softwares that KiCost can read, then exit.')
     parser.add_argument('-e', '--exclude',
                         nargs='+', type=str, default='',
                         metavar = 'dist',
@@ -126,6 +137,9 @@ def main():
                         default=HTML_RESPONSE_RETRIES,
                         metavar = 'num_retries',
                         help='Specify the number of attempts to retrieve part data from a website.')
+    parser.add_argument('--user',
+                        action='store_true',
+                        help='Start the user guide to run KiCost passing the file parameter give by "--input", all others parameters are ignored.')
 
 
     args = parser.parse_args()
@@ -145,6 +159,11 @@ def main():
     if args.show_dist_list:
         print('Distributor list:', *sorted(list(distributors.keys())))
         return
+    if args.show_eda_list:
+        #eda_names = [o[0] for o in inspect.getmembers(eda_tools_imports) if inspect.ismodule(o[1])]
+        #print('EDA supported list:', ', '.join(eda_names))
+        print('EDA supported list:', *sorted(list(eda_tool.keys())))
+        return
 
     # Set up spreadsheet output file.
     if args.output == None:
@@ -152,10 +171,10 @@ def main():
         if args.input != None:
             # Send output to spreadsheet with name of input file.
             if len(args.input)>1:
-            	# Compose a name with the multiple BOM input file names,
-            	# limiting to the first 5 caracheters of each name (avoid
-            	# huge names). THis is dynamic if the number of input
-            	# files passed.
+                # Compose a name with the multiple BOM input file names,
+                # limiting to the first 5 caracheters of each name (avoid
+                # huge names). This is dynamic if the number of input
+                # files passed.
                 args.output = '-'.join( [ os.path.splitext(args.input[i][:max(int(20/len(args.input)),5)])[0] for i in range(len(args.input))] ) + '.xlsx'
             else:
                 args.output = os.path.splitext(args.input[0])[0] + '.xlsx'
@@ -166,6 +185,16 @@ def main():
         # Output file was given. Make sure it has spreadsheet extension.
         args.output = os.path.splitext(args.output)[0] + '.xlsx'
 
+    # Call the KiCost interface to alredy run KiCost, this is just to use the
+    # saved user configurations of the graphical interface.
+    if args.user:
+        try:
+            kicost_gui_run([os.path.abspath(fileName) for fileName in args.input])
+        except NameError:
+            kicost_gui_notdependences()
+            #kicost_gui_run([os.path.abspath(fileName) for fileName in args.input])
+        return
+
     # Handle case where output is going into an existing spreadsheet file.
     if os.path.isfile(args.output):
         if not args.overwrite:
@@ -175,18 +204,24 @@ def main():
 
     # Set XML input source.
     if args.input == None:
-        # Get XML from the STDIN if no input file is given.
-        args.input = sys.stdin
+        try:
+            kicost_gui() # Use the user guide.
+        except NameError:
+            kicost_gui_notdependences()
+            #kicost_gui()
+        return
     else:
         # Otherwise get XML from the given file.
         for i in range(len(args.input)):
             # Set '.xml' as the default file extension, treating this exception
-            # allow (future) other files extension and formats.
-            if os.path.splitext(args.input[i])[1] == '':
-                args.input[i] += '.xml'
-            elif os.path.splitext(args.input[i])[1] == '.csv' or args.eda_tool[i] == 'csv' or args.eda_tool[i] == 'generic':
-                args.eda_tool = 'generic_csv'
-            args.input[i] = open(args.input[i])
+            # allow other files extension and formats.
+            try:
+                if os.path.splitext(args.input[i])[1] == '':
+                    args.input[i] += '.xml'
+                elif os.path.splitext(args.input[i])[1] == '.csv' or args.eda_tool[i] == 'csv':
+                    args.eda_tool = 'csv'
+            except IndexError:
+                pass
 
     # Set number of processes to use for web scraping.
     if args.serial:
@@ -209,3 +244,14 @@ if __name__ == '__main__':
     logger = logging.getLogger('kicost')
     logger.log(logging.DEBUG-2, 'Elapsed time: %f seconds', time.time() - start_time)
 
+
+###############################################################################
+# Additional functions
+###############################################################################
+
+def kicost_gui_notdependences():
+    print('You don\'t have the wxPython dependence to run the GUI interface. Run once of the follow commands in terminal to install them:')
+    print('pip install wxPython # For python 2.5')
+    print('pip3 install wxPython # For python 3.0+')
+    print('Or download from last version from <https://wxpython.org/>')
+    sys.exit(1)